# packages/api-types/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "JellyfinSuite Plugin API",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8096",
      "description": "Jellyfin server (default port)"
    }
  ],
  "security": [],
  "paths": {
...
    "/JellyfinSuite/FrameExport/{itemId}": {
      "get": {
        "tags": [
          "FrameExport"
        ],
        "operationId": "frameExport_getFrame",
        "summary": "Get Frame",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "frameIdx",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "positionMs",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "width",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "WebP frame",
            "content": {
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/JellyfinSuite/FrameExport/Prefetch/{itemId}": {
      "post": {
        "tags": [
          "FrameExport"
        ],
        "operationId": "frameExport_prefetch",
        "summary": "Prefetch",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrefetchRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
...
    "/JellyfinSuite/FrameExport/Generate": {
      "post": {
        "tags": [
          "FrameExport"
        ],
        "operationId": "frameExport_generate",
        "summary": "Generate",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Task accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/JellyfinSuite/FrameExport/TaskProgress": {
      "get": {
        "tags": [
          "FrameExport"
        ],
        "operationId": "frameExport_taskProgress",
        "summary": "Task Progress",
        "parameters": [
          {
            "name": "taskId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-Sent Events stream",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/TaskProgressDto"
                },
                "itemSchema": {
                  "type": "string",
                  "contentMediaType": "application/json",
                  "contentSchema": {
                    "$ref": "#/components/schemas/TaskProgressDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/JellyfinSuite/FrameExport/Result/{taskId}/{filename}": {
      "get": {
        "tags": [
          "FrameExport"
        ],
        "operationId": "frameExport_getResult",
        "summary": "Get Result",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result file",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/JellyfinSuite/FrameExport/Result/{taskId}": {
      "delete": {
        "tags": [
          "FrameExport"
        ],
        "operationId": "frameExport_deleteResult",
        "summary": "Delete Result",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/JellyfinSuite/FrameExport/Cancel/{taskId}": {
      "post": {
        "tags": [
          "FrameExport"
        ],
        "operationId": "frameExport_cancel",
        "summary": "Cancel",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancelled"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Already complete"
          }
        }
      }
    },
...
    "/JellyfinSuite/FrameExport/Debug": {
      "get": {
        "tags": [
          "FrameExport"
        ],
        "operationId": "frameExport_debug",
        "summary": "Debug",
        "responses": {
          "200": {
            "description": "Debug JSON"
          },
          "503": {
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/JellyfinSuite/FrameExport/QualityThresholds": {
      "get": {
        "tags": [
          "FrameExport"
        ],
        "operationId": "frameExport_getQualityThresholds",
        "summary": "Get Quality Thresholds",
        "responses": {
          "200": {
            "description": "Quality thresholds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualityThresholds"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "FrameExport"
        ],
        "operationId": "frameExport_setQualityThresholds",
        "summary": "Set Quality Thresholds",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QualityThresholds"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated thresholds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualityThresholds"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PrefetchRequest": {
        "type": "object",
        "properties": {
          "startFrameIdx": {
            "type": "integer",
            "format": "int32"
          },
          "beforeSeconds": {
            "type": "number",
            "format": "double"
          },
          "afterSeconds": {
            "type": "number",
            "format": "double"
          },
          "includeStart": {
            "type": "boolean"
          },
          "width": {
            "type": "integer",
            "format": "int32"
          },
          "positions": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "framePairs": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/FramePairDto"
            }
          }
        },
        "required": [
          "startFrameIdx",
          "beforeSeconds",
          "afterSeconds",
          "includeStart",
          "width"
        ]
      },
      "FramePairDto": {
        "type": "object",
        "properties": {
          "fiIdx": {
            "type": "integer",
            "format": "int64"
          },
          "posMs": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "fiIdx",
          "posMs"
        ]
      },
      "GenerateRequest": {
        "type": "object",
        "properties": {
          "itemId": {
            "type": "string",
            "format": "uuid"
          },
          "itemTitle": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "frames": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FrameReference"
            }
          },
          "params": {
            "$ref": "#/components/schemas/ExportParams"
          }
        },
        "required": [
          "itemId",
          "itemTitle",
          "type",
          "frames",
          "params"
        ]
      },
      "FrameReference": {
        "type": "object",
        "properties": {
          "frameIdx": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "positionMs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
...
      "GenerateResponse": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string"
          }
        },
        "required": [
          "taskId"
        ]
      },
      "TaskProgressDto": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "phase": {
            "type": "string"
          },
          "current": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "percent": {
            "type": "number",
            "format": "double"
          },
          "resultUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "fileSize": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "taskId",
          "status",
          "phase",
          "current",
          "total",
          "percent"
        ]
      },
      "QualityThresholds": {
        "type": "object",
        "properties": {
          "blackBrightnessVarMin": {
            "type": "number",
            "format": "double"
          },
          "whiteBrightnessVarMax": {
            "type": "number",
            "format": "double"
          },
          "blurLaplacianVarMin": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "blackBrightnessVarMin",
          "whiteBrightnessVarMax",
          "blurLaplacianVarMin"
        ]
      },
...
    }
  }
}