# packages/api-types/src/model/frameReference.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * JellyfinSuite Plugin API
 * OpenAPI spec version: 1.0.0
 */

export interface FrameReference {
  /** @nullable */
  frameIdx?: number | null;
  /** @nullable */
  positionMs?: number | null;
}